Merge Marlin 1.0.2 into BIBO2 source code
Attempt to merge Marlin 1.0.2 into the BIBO2 source code

BIBO2 source code is a modified version of the shaqfoo ditto print firmware.

This merge brings back the DUAL_X_CARRIAGE from marlin 1.0.2 to keep parity with Marlin repo so code changes are easier to follow. shaqfoo stripped the DUAL_X_CARRIAGE stuff out because they rightly believed that most printers do not have DUAL_X_CARRIAGE capabilities.

read more on the very old shaqfoo version of marlin
https://github.com/ShaqFoo/Marlin_Firmware_Deluxe_With_Ditto_Print-Auto_Leveling

most changes from the BIBO version of the firmware were maintained. some changes to the shaqfoo ditto printing were necessary to make sense of the code and intended functions. the biggest change was to steppers.h where the ditto printing if statement is a big question as to if it will work as rewritten.